+++ index.d.ts
@@ -6,29 +6,36 @@ type UpdateMetadata = MBTiles.UpdateMetadata
 type Bounds = MBTiles.Bounds
 type BBox = MBTiles.BBox
 type Center = MBTiles.Center
+type Schema = 'xyz' | 'tms' | 'quadkey'
+
+interface SchemaType {
+  xyz: Tile
+  tms: Tile
+  quadkey: string
+}
 
 /**
  * MBTiles
  */
-declare class MBTiles {
-  name: string
-  description: string
+declare class MBTilesStatic<T extends Tile | string> {
+  type: string
+  version: string
+  name?: string
   minzoom?: number
   maxzoom?: number
   format?: string
   bounds?: BBox
-  type?: string
-  version?: string
   center?: Center
+  description?: string
   attribution?: string
   url?: string
 
-  constructor(uri: string)
+  constructor(uri: string, schema?: Schema)
 
   /**
    * Save buffer data to individual Tile
    */
-  save(tile: Tile, image: Buffer): Promise<boolean>
+  save(tile: T, image: Buffer): Promise<boolean>
 
   /**
    * Retrieves Metadata from MBTiles
@@ -38,12 +45,12 @@ declare class MBTiles {
   /**
    * Delete individual Tile
    */
-  delete(tile: Tile): Promise<boolean>
+  delete(tile: T): Promise<boolean>
 
   /**
    * Count the amount of Tiles
    */
-  count(tiles?: Tile[]): Promise<number>
+  count(tiles?: T[]): Promise<number>
 
 
   /**
@@ -54,12 +61,12 @@ declare class MBTiles {
   /**
    * Finds all Tile unique hashes
    */
-  findAll(tiles?: Tile[]): Promise<Tile[]>
+  findAll(tiles?: T[]): Promise<T[]>
 
   /**
    * Finds one Tile and returns Buffer
    */
-  findOne(tile: Tile): Promise<Buffer>
+  findOne(tile: T): Promise<Buffer>
 
   /**
    * Build SQL tables
@@ -74,12 +81,12 @@ declare class MBTiles {
   /**
    * Creates hash from a single Tile
    */
-  hash(tile: Tile): number
+  hash(tile: T): number
 
   /**
    * Creates a hash table for all tiles
    */
-  hashes(tiles?: Tile[]): Promise<MBTiles.Hashes>
+  hashes(tiles?: T[]): Promise<MBTiles.Hashes>
 
   /**
    * Retrieves Minimum Zoom level
@@ -119,10 +126,7 @@ declare namespace MBTiles {
   export type MultiPolygon = GeoJSON.Feature<GeoJSON.MultiPolygon>
   export type MultiPolygons = GeoJSON.FeatureCollection<GeoJSON.MultiPolygon>
   export type Bounds = BBox | BBox[] | Polygon | Polygons | MultiPolygon | MultiPolygons
-
-  export interface Hashes {
-    [key: number]: boolean
-  }
+  export type Hashes = Set<number>
 
   export interface Metadata extends BaseMetadata {
     bounds?: BBox
@@ -145,4 +149,11 @@ declare namespace MBTiles {
     url?: string
   }
 }
+
+interface MBTiles<T> {
+  new (uri: string, schema: 'quadkey'): MBTilesStatic<string>;
+  new (uri: string, schema?: 'xyz' | 'tms'): MBTilesStatic<Tile>;
+}
+
+declare const MBTiles: MBTiles<Tile>;
 export = MBTiles